Refactored `shared.py`
Created a hierarchy of classes for accessing file attributes available
on different platforms. Created `BaseFile` class as the base class for
all other file-related class, `WindowsFile` for accessing file attributes
available on Windows, `PosixFile` for accessing the same on Mac and Linux,
and `File` class to create an instance of the File object corresponding to
the platform of operation.