Update ConditionHandler class
Updated the `__init__` method to accept None for the `conditions`
parameter and store it unaltered in the associated attribute to signify
that no conditions are to be applied for evaluation.

Also updated the `__call__` method based on the changes mentioned above
to return True if no conditions were explicitly specified to include all
the query records.