Refactor a parser method in ConditionParser class
Refactor the `_parse_relational` method adding an exceptional case
for a successful operation where the operator is Equals (==) or
Not Equals (!=) and one of the operand is None allowing for comparsion
with query fields in case it returns a None. Updated the conditional
statement to avoid raising an error in the above mentioned case.